requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0

# Computer vision
opencv-python>=4.8.0
Pillow>=9.5.0

# Face detection
facenet-pytorch>=2.5.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Machine learning metrics
scikit-learn>=1.3.0

# Web interface
streamlit>=1.25.0

# Utilities
tqdm>=4.65.0

# Optional: For CUDA support
# torch with CUDA: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118